Extract fetch stats and refactor some
