+++ src/App.tsx
@@ -11,9 +11,9 @@ import PrivateRoutesController from './routes/PrivateRoutesController'
 // Components
 import Navbar from './components/navbar/Navbar'
 import Sidebar from './components/sidebar/Sidebar'
-import LoggingLoader from './components/loader/LoggingLoader'
 import PublicRoute from './components/route/PublicRoute'
 import AuthPage from './pages/auth/AuthPage'
+import BasicLoader from './components/loader/BasicLoader'
 
 // State
 import { userState } from './global-state/miscState'
@@ -24,9 +24,9 @@ import client from './api/client'
 // Hooks
 import useLoadActiveListData from './hooks/useLoadActiveListData'
 import useFetchCategories from './hooks/useFetchCategories'
-import BasicLoader from './components/loader/BasicLoader'
 import useFetchItems from './hooks/useFetchItems'
-import useLoadHistoryLists from './hooks/useLoadHistoryLists'
+import useLoadHistoryLists from './hooks/useFetchHistoryLists'
+import useFetchStats from './hooks/useFetchStats'
 
 /**
  * Main app component
@@ -45,6 +45,7 @@ const App: React.FC = () => {
     const fetchCategories = useFetchCategories()
     const fetchItems = useFetchItems()
     const fetchShopListHistory = useLoadHistoryLists()
+    const fetchStats = useFetchStats()
 
     useEffect(() => {
         if (location.search.length > 0) {
@@ -77,6 +78,7 @@ const App: React.FC = () => {
         fetchItems()
         initialActiveShopListData()
         fetchShopListHistory()
+        fetchStats()
     }
 
     useEffect(() => {
