+++ src/global-state/statisticsState.tsx
@@ -3,15 +3,19 @@ import { atom } from 'recoil'
 export const statisticsState = atom({
     key: 'statistics',
     default: {
-        month: {
-            itemsByMonth: [],
-            categoriesByMonth: [],
-            quantityByDay: [],
-        },
-        year: {
-            itemsByYear: [],
-            categoriesByYear: [],
-            quantityByMonth: [],
+        loading: false,
+        noStats: false,
+        stats: {
+            month: {
+                itemsByMonth: [],
+                categoriesByMonth: [],
+                quantityByDay: [],
+            },
+            year: {
+                itemsByYear: [],
+                categoriesByYear: [],
+                quantityByMonth: [],
+            },
         },
     },
 })
