+++ src/global-state/statsState.tsx
@@ -1,0 +1,9 @@
+import { atom } from 'recoil'
+
+export const statsState = atom({
+    key: 'statsState',
+    default: {
+        loading: false,
+        stats: {},
+    },
+})
